Split bbinternal package and move code
pkg/bb/bbinternal -> pkg/bb (code that modifies AST, creates main.go) &
pkg/stdgobb/bbpkg (code that turns user input into a *bb.Package list)

pkg/bb -> pkg/stdgobb (standard Go bb: stuff that handles modules & GOPATH)
