+++ src/pkg/bb/BUILD.bazel
@@ -1,21 +1,14 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_embed_data", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "bb",
-    srcs = [
-        "bb.go",
-        "bbmain_src.go",
-        "bbregister_src.go",
-        "generate.go",
-    ],
+    srcs = ["bb.go"],
     importpath = "github.com/u-root/gobusybox/src/pkg/bb",
     visibility = ["//visibility:public"],
     deps = [
-        "//src/pkg/bb/bbinternal",
-        "//src/pkg/golang",
-        "@com_github_google_goterm//term",
         "@com_github_u_root_uio//cp",
         "@org_golang_x_tools//go/ast/astutil",
         "@org_golang_x_tools//go/packages",
+        "@org_golang_x_tools//imports",
     ],
 )
