+++ src/pkg/bb/bb.go
@@ -2,612 +2,528 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package bb builds one busybox-like binary out of many Go command sources.
+// Package bb provides APIs for rewriting Go source files to
 //
-// This allows you to take two Go commands, such as Go implementations of `sl`
-// and `cowsay` and compile them into one binary, callable like `./bb sl` and
-// `./bb cowsay`. Which command is invoked is determined by `argv[0]` or
-// `argv[1]` if `argv[0]` is not recognized.
+// is the internal API for both bazel and standard Go
+// busybox builds.
 //
-// Under the hood, bb implements a Go source-to-source transformation on pure
-// Go code. This AST transformation does the following:
-//
-//   - Takes a Go command's source files and rewrites them into Go package files
-//     without global side effects.
-//   - Writes a `main.go` file with a `main()` that calls into the appropriate Go
-//     command package based on `argv[0]`.
-//
-// Principally, the AST transformation moves all global side-effects into
-// callable package functions. E.g. `main` becomes `registeredMain`, each
-// `init` becomes `initN`, and global variable assignments are moved into their
-// own `initN`.
+// It contains exported functions that are not for user consumption and not
+// stable
 package bb
 
 import (
+	"bytes"
 	"fmt"
-	"io"
+	"go/ast"
+	"go/format"
+	"go/parser"
+	"go/token"
+	"go/types"
 	"io/ioutil"
-	"log"
 	"os"
 	"path"
 	"path/filepath"
 	"regexp"
+	"sort"
+	"strconv"
 	"strings"
 
-	"github.com/google/goterm/term"
+	"github.com/u-root/uio/cp"
 	"golang.org/x/tools/go/ast/astutil"
 	"golang.org/x/tools/go/packages"
-
-	"github.com/u-root/gobusybox/src/pkg/bb/bbinternal"
-	"github.com/u-root/gobusybox/src/pkg/golang"
-	"github.com/u-root/uio/cp"
+	"golang.org/x/tools/imports"
 )
 
-func listStrings(m map[string]struct{}) []string {
-	var l []string
-	for k := range m {
-		l = append(l, k)
-	}
-	return l
-}
+// The Go spec defines the following grammar:
+//   identifier = letter { letter | unicode_digit } .
+// See also https://golang.org/ref/spec#Identifiers
+var pnameRegex = regexp.MustCompile("[^a-zA-Z0-9_]+")
 
-func checkDuplicate(cmds []*bbinternal.Package) error {
-	seen := make(map[string]string)
-	for _, cmd := range cmds {
-		if path, ok := seen[cmd.Name]; ok {
-			return fmt.Errorf("failed to build with bb: found duplicate command %s (%s and %s)", cmd.Name, path, cmd.Pkg.PkgPath)
-		}
-		seen[cmd.Name] = cmd.Pkg.PkgPath
-	}
-	return nil
-}
-
-// Opts are the arguments to BuildBusybox.
-type Opts struct {
-	// Env are the environment variables used in Go compilation and package
-	// discovery.
-	Env golang.Environ
-
-	// GenSrcDir is an empty directory to generate the busybox source code
-	// in.
-	//
-	// If GenSrcDir has children, BuildBusybox will return an error. If
-	// GenSrcDir does not exist, it will be created. If no GenSrcDir is
-	// given, a temporary directory will be generated. The generated
-	// directory will be deleted if compilation succeeds.
-	//
-	// In GOPATH mode, GOPATH=GenSrcDir for compilation.
-	GenSrcDir string
-
-	// CommandPaths is a list of file system directories containing Go
-	// commands, or Go import paths.
-	CommandPaths []string
-
-	// BinaryPath is the file to write the binary to.
-	BinaryPath string
-
-	// GoBuildOpts is configuration for the `go build` command that
-	// compiles the busybox binary.
-	GoBuildOpts *golang.BuildOpts
-
-	// AllowMixedMode allows mixed mode (module / non-module) compilation.
-	//
-	// If this is done with GO111MODULE=on,
-	AllowMixedMode bool
-}
-
-// BuildBusybox builds a busybox of many Go commands. opts contains both the
-// commands to build and other options.
+// ParseAST parses the given files for a package named name.
 //
-// For documentation on how this works, please refer to the README at the top
-// of the repository.
-func BuildBusybox(opts *Opts) (nerr error) {
-	if opts == nil {
-		return fmt.Errorf("no options given for busybox build")
-	} else if err := opts.Env.Valid(); err != nil {
-		return err
-	}
-
-	var tmpDir string
-	if opts.GenSrcDir != "" {
-		var relTmpDir string
-		dirents, err := ioutil.ReadDir(opts.GenSrcDir)
-		if os.IsNotExist(err) {
-			if err := os.MkdirAll(opts.GenSrcDir, 0700); err != nil {
-				return fmt.Errorf("could not create busybox generated source directory: %w", err)
-			}
-			relTmpDir = opts.GenSrcDir
+// Only files with a matching package statement will be part of the AST
+// returned.
+func ParseAST(name string, files []string) (*token.FileSet, []*ast.File, []string, error) {
+	fset := token.NewFileSet()
+	astFiles := make(map[string]*ast.File)
+	for _, path := range files {
+		if src, err := parser.ParseFile(fset, path, nil, parser.ParseComments); err == nil && src.Name.Name == name {
+			astFiles[path] = src
 		} else if err != nil {
-			return fmt.Errorf("could not read busybox generated source directory: %w", err)
-		} else if len(dirents) > 0 {
-			return fmt.Errorf("busybox generated source directory is not an empty directory")
-		} else {
-			relTmpDir = opts.GenSrcDir
-		}
-		absDir, err := filepath.Abs(relTmpDir)
-		if err != nil {
-			return fmt.Errorf("busybox gen src dir %s could not be made absolute: %v", relTmpDir, err)
-		}
-		tmpDir = absDir
-	} else {
-		var err error
-		tmpDir, err = ioutil.TempDir("", "bb-")
-		if err != nil {
-			return err
-		}
-		defer func() {
-			if nerr != nil {
-				log.Printf("Preserving bb generated source directory at %s due to error", tmpDir)
-			} else {
-				os.RemoveAll(tmpDir)
-			}
-		}()
-	}
-
-	bbDir := filepath.Join(tmpDir, "src/bb.u-root.com/bb")
-	if err := os.MkdirAll(bbDir, 0700); err != nil {
-		return err
-	}
-	pkgDir := filepath.Join(tmpDir, "src")
-
-	// Ask go about all the commands in one batch for dependency caching.
-	cmds, err := bbinternal.NewPackages(opts.Env, opts.CommandPaths...)
-	if err != nil {
-		return fmt.Errorf("finding packages failed: %v", err)
-	}
-	if len(cmds) == 0 {
-		return fmt.Errorf("no valid commands given")
-	}
-
-	// Collect all packages that we need to actually re-write.
-	if err := checkDuplicate(cmds); err != nil {
-		return err
-	}
-
-	modules := make(map[string]struct{})
-	var numNoModule int
-	for _, cmd := range cmds {
-		if cmd.Pkg.Module != nil {
-			modules[cmd.Pkg.Module.Path] = struct{}{}
-		} else {
-			numNoModule++
+			return nil, nil, nil, fmt.Errorf("failed to parse AST in file %q: %v", path, err)
 		}
 	}
-	if !opts.AllowMixedMode && len(modules) > 0 && numNoModule > 0 {
-		return fmt.Errorf("busybox does not support mixed module/non-module compilation -- commands contain main modules %v", strings.Join(listStrings(modules), ", "))
+
+	// Did we parse anything?
+	if len(astFiles) == 0 {
+		return nil, nil, nil, fmt.Errorf("no valid `main` package files found in %v", files)
 	}
 
-	// List of packages to import in the real main file.
-	var bbImports []string
-	// Rewrite commands to packages.
-	for _, cmd := range cmds {
-		destination := filepath.Join(pkgDir, cmd.Pkg.PkgPath)
+	// The order of types.Info.InitOrder depends on this list of files
+	// always being passed to conf.Check in the same order.
+	sort.Strings(files)
 
-		if err := cmd.Rewrite(destination, "bb.u-root.com/bb/pkg/bbmain"); err != nil {
-			return fmt.Errorf("rewriting command %q failed: %v", cmd.Pkg.PkgPath, err)
-		}
-		bbImports = append(bbImports, cmd.Pkg.PkgPath)
-	}
-
-	// Collect and write dependencies into pkgDir.
-	if err := dealWithDeps(opts.Env, bbDir, tmpDir, pkgDir, cmds); err != nil {
-		return fmt.Errorf("collecting and putting dependencies in place failed: %v", err)
-	}
-
-	if err := writeBBMain(bbDir, tmpDir, bbImports); err != nil {
-		return fmt.Errorf("failed to write main.go: %v", err)
-	}
-
-	// Compile bb.
-	if opts.Env.GO111MODULE == "off" || numNoModule > 0 {
-		opts.Env.GOPATH = tmpDir
-	}
-	if err := opts.Env.BuildDir(bbDir, opts.BinaryPath, opts.GoBuildOpts); err != nil {
-		if opts.Env.GO111MODULE == "off" || numNoModule > 0 {
-			return &ErrGopathBuild{
-				CmdDir: bbDir,
-				GOPATH: tmpDir,
-				Err:    err,
-			}
-		} else {
-			return &ErrModuleBuild{
-				CmdDir: bbDir,
-				Err:    err,
-			}
+	sortedFiles := make([]*ast.File, 0, len(astFiles))
+	parsedFiles := make([]string, 0, len(astFiles))
+	for _, name := range files {
+		if f, ok := astFiles[name]; ok {
+			sortedFiles = append(sortedFiles, f)
+			parsedFiles = append(parsedFiles, name)
 		}
 	}
-	return nil
+	return fset, sortedFiles, parsedFiles, nil
 }
 
-// ErrModuleBuild is returned for a go build failure when modules were enabled.
-type ErrModuleBuild struct {
-	CmdDir string
-	Err    error
-}
-
-// Unwrap implements error.Unwrap.
-func (e *ErrModuleBuild) Unwrap() error {
-	return e.Err
-}
-
-// Error implements error.Error.
-func (e *ErrModuleBuild) Error() string {
-	return fmt.Sprintf("go build with modules failed: %v", e.Err)
-}
-
-// ErrGopathBuild is returned for a go build failure when modules were disabled.
-type ErrGopathBuild struct {
-	CmdDir string
-	GOPATH string
-	Err    error
-}
-
-// Unwrap implements error.Unwrap.
-func (e *ErrGopathBuild) Unwrap() error {
-	return e.Err
-}
-
-// Error implements error.Error.
-func (e *ErrGopathBuild) Error() string {
-	return fmt.Sprintf("non-module go build failed: %v", e.Err)
-}
-
-// writeBBMain writes $TMPDIR/src/bb.u-root.com/bb/pkg/bbmain/register.go and
-// $TMPDIR/src/bb.u-root.com/bb/main.go.
+// CreateBBMainSource creates a bb Go command main.go that imports all given
+// pkgs and writes the command to destDir.
 //
-// They are taken from ./bbmain/register.go and ./bbmain/cmd/main.go, but they
-// do not retain their original import paths because the main command must be
-// in a module that doesn't conflict with any bb commands. If one were to
-// compile github.com/u-root/gobusybox/src/cmd/* into a busybox, we'd have
-// problems -- the src/go.mod would conflict with our generated go.mod, and
-// it'd be complicated to merge them. So they are transplanted into the
-// bb.u-root.com/bb module.
-func writeBBMain(bbDir, tmpDir string, bbImports []string) error {
-	if err := os.MkdirAll(filepath.Join(bbDir, "pkg/bbmain"), 0755); err != nil {
-		return err
-	}
-	if err := ioutil.WriteFile(filepath.Join(bbDir, "pkg/bbmain/register.go"), bbRegisterSource, 0755); err != nil {
-		return err
-	}
-	if err := ioutil.WriteFile(filepath.Join(bbDir, "main.go"), bbMainSource, 0755); err != nil {
-		return err
+// fset and files must be parsed bb template main.go, usually ./bbmain/cmd/main.go.
+func CreateBBMainSource(fset *token.FileSet, files []*ast.File, pkgs []string, destDir string) error {
+	if len(files) != 1 {
+		return fmt.Errorf("bb cmd template is supposed to only have one file")
 	}
 
-	bbFset, bbFiles, _, err := bbinternal.ParseAST("main", []string{filepath.Join(bbDir, "main.go")})
-	if err != nil {
-		return err
+	for _, pkg := range pkgs {
+		astutil.AddNamedImport(fset, files[0], "_", pkg)
 	}
-	if len(bbFiles) == 0 {
-		return fmt.Errorf("bb package not found")
-	}
-
-	// Fix the import path for bbmain, since we wrote bbmain/register.go into bbDir above.
-	if !astutil.RewriteImport(bbFset, bbFiles[0], "github.com/u-root/gobusybox/src/pkg/bb/bbmain", "bb.u-root.com/bb/pkg/bbmain") {
-		return fmt.Errorf("could not rewrite import")
-	}
-
-	// Create bb main.go.
-	if err := bbinternal.CreateBBMainSource(bbFset, bbFiles, bbImports, bbDir); err != nil {
-		return fmt.Errorf("creating bb main.go file failed: %v", err)
-	}
-	return nil
+	return writeFiles(destDir, fset, files)
 }
 
-func isReplacedModuleLocal(m *packages.Module) bool {
-	// From "replace directive": https://golang.org/ref/mod#go
+// Package is a Go package.
+type Package struct {
+	// Name is the executable command name.
 	//
-	//   If the path on the right side of the arrow is an absolute or
-	//   relative path (beginning with ./ or ../), it is interpreted as the
-	//   local file path to the replacement module root directory, which
-	//   must contain a go.mod file. The replacement version must be
-	//   omitted in this case.
-	return strings.HasPrefix(m.Path, "./") || strings.HasPrefix(m.Path, "../") || strings.HasPrefix(m.Path, "/")
-}
+	// In the standard Go tool chain, this is usually the base name of the
+	// directory containing its source files.
+	Name string
 
-// localModules finds all modules that are local, copies their go.mod in the
-// right place, and raises an error if any modules have conflicting replace
-// directives.
-func localModules(pkgDir, bbDir string, mainPkgs []*bbinternal.Package) (map[string]*packages.Module, error) {
-	copyGoMod := func(mod *packages.Module) error {
-		if mod == nil {
-			return nil
-		}
+	// Pkg is the actual data about the package.
+	Pkg *packages.Package
 
-		if err := os.MkdirAll(filepath.Join(pkgDir, mod.Path), 0755); os.IsExist(err) {
-			return nil
-		} else if err != nil {
-			return err
-		}
+	// initCount keeps track of what the next init's index should be.
+	initCount uint
 
-		// Use the module file for all outside dependencies.
-		if err := cp.Copy(mod.GoMod, filepath.Join(pkgDir, mod.Path, "go.mod")); err != nil {
-			return err
-		}
+	// mainFuncName is the name for the renamed main().
+	mainFuncName string
 
-		// As of Go 1.16, the Go build system expects an accurate
-		// go.sum in the main module directory. We build it by
-		// concatenating all constituent go.sums.
-		//
-		// If it doesn't exist, that's okay!
-		gosum := filepath.Join(filepath.Dir(mod.GoMod), "go.sum")
-		if err := cp.Copy(gosum, filepath.Join(pkgDir, mod.Path, "go.sum")); os.IsNotExist(err) {
-			// Modules without dependencies don't have or need a go.sum.
-			return nil
-		} else if err != nil {
-			return err
-		}
+	// init is the cmd.Init function that calls all other InitXs in the
+	// right order.
+	init *ast.FuncDecl
 
-		gosumf, err := os.Open(gosum)
-		if err != nil {
-			return err
-		}
-		defer gosumf.Close()
-		f, err := os.OpenFile(filepath.Join(bbDir, "go.sum"), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0755)
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		_, err = io.Copy(f, gosumf)
-		return err
-	}
-
-	type localModule struct {
-		m          *packages.Module
-		provenance string
-	}
-	localModules := make(map[string]*localModule)
-	// Find all top-level modules.
-	for _, p := range mainPkgs {
-		if p.Pkg.Module != nil {
-			if _, ok := localModules[p.Pkg.Module.Path]; !ok {
-				localModules[p.Pkg.Module.Path] = &localModule{
-					m:          p.Pkg.Module,
-					provenance: fmt.Sprintf("your request to compile %s from %s", p.Pkg.Module.Path, p.Pkg.Module.Dir),
-				}
-
-				if err := copyGoMod(p.Pkg.Module); err != nil {
-					return nil, fmt.Errorf("failed to copy go.mod for %s: %v", p.Pkg.Module.Path, err)
-				}
-			}
-		}
-	}
-
-	for _, p := range mainPkgs {
-		replacedModules := locallyReplacedModules(p.Pkg)
-		for modPath, module := range replacedModules {
-			if original, ok := localModules[modPath]; ok {
-				// Is this module different from one that a
-				// previous definition provided?
-				//
-				// This only looks for 2 conflicting *local* module definitions.
-				if original.m.Dir != module.Dir {
-					return nil, fmt.Errorf("two conflicting versions of module %s have been requested; one from %s, the other from %s's go.mod",
-						modPath, original.provenance, p.Pkg.Module.Path)
-				}
-			} else {
-				localModules[modPath] = &localModule{
-					m:          module,
-					provenance: fmt.Sprintf("%s's go.mod (%s)", p.Pkg.Module.Path, p.Pkg.Module.GoMod),
-				}
-
-				if err := copyGoMod(module); err != nil {
-					return nil, fmt.Errorf("failed to copy go.mod for %s: %v", p.Pkg.Module.Path, err)
-				}
-			}
-		}
-	}
-
-	// Look for conflicts between remote and local modules.
+	// initAssigns is a map of assignment expression -> InitN function call
+	// statement.
 	//
-	// E.g. if u-bmc depends on u-root, but we are also compiling u-root locally.
-	var conflict bool
-	for _, mainPkg := range mainPkgs {
-		packages.Visit([]*packages.Package{mainPkg.Pkg}, nil, func(p *packages.Package) {
-			if p.Module == nil {
-				return
-			}
-			if l, ok := localModules[p.Module.Path]; ok && l.m.Dir != p.Module.Dir {
-				fmt.Fprintln(os.Stderr, "")
-				log.Printf("Conflicting module dependencies on %s:", p.Module.Path)
-				log.Printf("  %s uses %s", mainPkg.Pkg.Module.Path, moduleIdentifier(p.Module))
-				log.Printf("  %s uses %s", l.provenance, moduleIdentifier(l.m))
-				replacePath, err := filepath.Rel(mainPkg.Pkg.Module.Dir, l.m.Dir)
-				if err != nil {
-					replacePath = l.m.Dir
-				}
-				fmt.Fprintln(os.Stderr, "")
-				log.Printf("%s: add `replace %s => %s` to %s", term.Bold("Suggestion to resolve"), p.Module.Path, replacePath, mainPkg.Pkg.Module.GoMod)
-				fmt.Fprintln(os.Stderr, "")
-				conflict = true
-			}
-		})
-	}
-	if conflict {
-		return nil, fmt.Errorf("conflicting module dependencies found")
-	}
-
-	modules := make(map[string]*packages.Module)
-	for modPath, mod := range localModules {
-		modules[modPath] = mod.m
-	}
-	return modules, nil
+	// That InitN should contain the assignment statement for the
+	// appropriate assignment expression.
+	//
+	// types.Info.InitOrder keeps track of Initializations by Lhs name and
+	// Rhs ast.Expr.  We reparent the Rhs in assignment statements in InitN
+	// functions, so we use the Rhs as an easy key here.
+	// types.Info.InitOrder + initAssigns can then easily be used to derive
+	// the order of Stmts in the "real" init.
+	//
+	// The key Expr must also be the AssignStmt.Rhs[0].
+	initAssigns map[ast.Expr]ast.Stmt
 }
 
-func moduleIdentifier(m *packages.Module) string {
-	if m.Replace != nil && isReplacedModuleLocal(m.Replace) {
-		return fmt.Sprintf("directory %s", m.Replace.Path)
+// NewPackage creates a new Package based on an existing packages.Package.
+func NewPackage(name string, p *packages.Package) *Package {
+	pp := &Package{
+		// Name is the executable name.
+		Name:        path.Base(name),
+		Pkg:         p,
+		initAssigns: make(map[ast.Expr]ast.Stmt),
 	}
-	return fmt.Sprintf("version %s", m.Version)
+
+	pp.mainFuncName = pp.newFunctionName("registeredMain")
+
+	// This Init will hold calls to all other InitXs.
+	pp.init = &ast.FuncDecl{
+		Name: ast.NewIdent(pp.newFunctionName("registeredInit")),
+		Type: &ast.FuncType{
+			Params:  &ast.FieldList{},
+			Results: nil,
+		},
+		Body: &ast.BlockStmt{},
+	}
+	return pp
 }
 
-// dealWithDeps tries to suss out local files that need to be in the tree.
+func (p *Package) nextInit(addToCallList bool) *ast.Ident {
+	nextInitName := fmt.Sprintf("busyboxInit%d", p.initCount)
+	for p.funcNameTaken(nextInitName) {
+		p.initCount++
+		nextInitName = fmt.Sprintf("busyboxInit%d", p.initCount)
+	}
+	i := ast.NewIdent(nextInitName)
+	if addToCallList {
+		p.init.Body.List = append(p.init.Body.List, &ast.ExprStmt{X: &ast.CallExpr{Fun: i}})
+	}
+	p.initCount++
+	return i
+}
+
+// importName finds the package path to import, given the go/types pkg path.
 //
-// It helps to have read https://golang.org/ref/mod when editing this function.
-func dealWithDeps(env golang.Environ, bbDir, tmpDir, pkgDir string, mainPkgs []*bbinternal.Package) error {
-	// Module-enabled Go programs resolve their dependencies in one of two ways:
+// E.g. go/types uses the fully vendored name of a package, such as
+// github.com/u-root/u-root/vendor/golang.org/x/sys/unix. importName returns
+// the name that should appear in the import statement for this package, which
+// is golang.org/x/sys/unix.
+//
+// Since the only way this happens is through an implicit import, we know that
+// somewhere in the dependency tree this package must exist, so we visit all
+// dependencies looking for the go/types package path looking for a valid
+// package import path statement.
+func importName(p *packages.Package, typePkgPath string) string {
+	var importName string
+	// Go through all dependent packages.
+	packages.Visit([]*packages.Package{p}, func(p *packages.Package) bool {
+		// Yeah, packages.Visit already goes through all imports -- but
+		// it does not give us the index of the p.Imports map, which is
+		// the "import paths appearing in the package's Go source
+		// files".
+		for name, pkg := range p.Imports {
+			if pkg.PkgPath == typePkgPath {
+				importName = name
+				return false
+			}
+		}
+		return true
+	}, nil)
+	if len(importName) > 0 {
+		return importName
+	}
+	// It doesn't appear. We'll go import it.
+	return typePkgPath
+}
+
+// pkgImportNameTaken checks whether name would conflict with any existing
+// imports in f or variable/const/func declarations in p.
+//
+// Import statements may conflict with import statements in other files in
+// the same package.
+func (p *Package) pkgImportNameTaken(name string, f *ast.File) bool {
+	// package scope is all variable, const, and func names
+	if p.Pkg.Types.Scope().Lookup(name) != nil {
+		return true
+	}
+
+	// file scope is imports. Only imports from this file can conflict.
+	// Imports in other files have no effect.
+	if p.Pkg.TypesInfo.Scopes[f].Lookup(name) != nil {
+		return true
+	}
+	return false
+}
+
+// funcNameTaken checks whether name would conflict with any
+// import/variable/const/func declarations in all of p.
+//
+// Variable/const/func names may not conflict with import statements in
+// other files of the same package!
+func (p *Package) funcNameTaken(name string) bool {
+	// package scope is all variable, const, and func names
+	if p.Pkg.Types.Scope().Lookup(name) != nil {
+		return true
+	}
+
+	// file scope is all imports
+	for _, file := range p.Pkg.Syntax {
+		if p.Pkg.TypesInfo.Scopes[file].Lookup(name) != nil {
+			return true
+		}
+	}
+	return false
+}
+
+// newFunctionName returns an unused function name in p with the prefix name.
+func (p *Package) newFunctionName(name string) string {
+	var i int
+	proposed := name
+	for p.funcNameTaken(proposed) {
+		proposed = fmt.Sprintf("%s%d", name, i)
+		i++
+	}
+	return proposed
+}
+
+// newImportName returns an unused import name in f/p with the prefix name.
+func (p *Package) newImportName(name string, f *ast.File) string {
+	var i int
+	proposed := name
+	for p.pkgImportNameTaken(proposed, f) {
+		proposed = fmt.Sprintf("%s%d", name, i)
+		i++
+	}
+	return proposed
+}
+
+func (p *Package) rewriteFile(f *ast.File) bool {
+	hasMain := false
+
+	// Change the package name declaration from main to the command's name.
+	// Remove all non-alphanumeric characters except for underscore and ensure
+	// starting with a letter. There are more valid identifiers though.
+	f.Name.Name = "bb" + pnameRegex.ReplaceAllString(p.Name, "")
+
+	// Map of fully qualified package name -> imported alias in the file.
+	importAliases := make(map[string]string)
+	unaliasedImports := make(map[string]struct{})
+	for _, impt := range f.Imports {
+		importPath, err := strconv.Unquote(impt.Path.Value)
+		if err != nil {
+			panic(err)
+		}
+
+		if impt.Name != nil {
+			importAliases[importPath] = impt.Name.Name
+		} else {
+			// We do not record the name of the package, because we
+			// do not know it. However, `qualifier` will know it
+			// because it's passed in.
+			unaliasedImports[importPath] = struct{}{}
+		}
+	}
+
+	// When the types.TypeString function translates package names, it uses
+	// this function to map fully qualified package paths to a local alias,
+	// if it exists.
+	qualifier := func(pkg *types.Package) string {
+		// pkg.Path() = fully vendored package name.
+		// importPath = package name as it appears in `import` statement.
+		importPath := importName(p.Pkg, pkg.Path())
+
+		if name, ok := importAliases[importPath]; ok {
+			return name
+		}
+		if _, ok := unaliasedImports[importPath]; ok {
+			// The package name is NOT the last component of its path.
+			return pkg.Name()
+		}
+		// When referring to self, don't use any package name.
+		if pkg == p.Pkg.Types {
+			return ""
+		}
+
+		// This type is not imported in this file yet.
+		//
+		// This typically happens when a derived global import uses a
+		// type that was previously only implicitly used.
+		//
+		// E.g. if we call func Foo() *log.Logger like this:
+		//
+		//   var l = pkg.Foo()
+		//
+		// Then it's possible the `log` package was not referred to at
+		// all previously, and we now need to add an import for log.
+		importAlias := p.newImportName(pkg.Name(), f)
+		astutil.AddNamedImport(p.Pkg.Fset, f, importAlias, importPath)
+		// Make sure we do not add this import twice.
+		importAliases[importPath] = importAlias
+
+		return importAlias
+	}
+
+	for _, decl := range f.Decls {
+		switch d := decl.(type) {
+		case *ast.GenDecl:
+			// We only care about vars.
+			if d.Tok != token.VAR {
+				break
+			}
+			for _, spec := range d.Specs {
+				s := spec.(*ast.ValueSpec)
+				if s.Values == nil {
+					continue
+				}
+
+				// For each assignment, create a new init
+				// function, and place it in the same file.
+				for i, name := range s.Names {
+					varInit := &ast.FuncDecl{
+						Name: p.nextInit(false),
+						Type: &ast.FuncType{
+							Params:  &ast.FieldList{},
+							Results: nil,
+						},
+						Body: &ast.BlockStmt{
+							List: []ast.Stmt{
+								&ast.AssignStmt{
+									Lhs: []ast.Expr{name},
+									Tok: token.ASSIGN,
+									Rhs: []ast.Expr{s.Values[i]},
+								},
+							},
+						},
+					}
+					// Add a call to the new init func to
+					// this map, so they can be added to
+					// Init0() in the correct init order
+					// later.
+					p.initAssigns[s.Values[i]] = &ast.ExprStmt{X: &ast.CallExpr{Fun: varInit.Name}}
+					f.Decls = append(f.Decls, varInit)
+				}
+
+				// Add the type of the expression to the global
+				// declaration instead.
+				if s.Type == nil {
+					typ := p.Pkg.TypesInfo.Types[s.Values[0]]
+					s.Type = ast.NewIdent(types.TypeString(typ.Type, qualifier))
+				}
+				s.Values = nil
+			}
+
+		case *ast.FuncDecl:
+			if d.Recv == nil && d.Name.Name == "main" {
+				d.Name.Name = p.mainFuncName
+				hasMain = true
+			}
+			if d.Recv == nil && d.Name.Name == "init" {
+				d.Name = p.nextInit(true)
+			}
+		}
+	}
+
+	// Now we change any import names attached to package declarations. We
+	// just upcase it for now; it makes it easy to look in bbsh for things
+	// we changed, e.g. grep -r bbsh Import is useful.
 	//
-	// - locally, if the dependency is *in* the module or there is a local replace directive
-	// - remotely, if not local
-	//
-	// I.e. if the module is github.com/u-root/u-root,
-	//
-	// - local: github.com/u-root/u-root/pkg/uio
-	// - remote: github.com/hugelgupf/p9/p9
-	// - also local: a remote module, with a local replace rule
-	//
-	// For local dependencies, we copy all dependency packages' files over,
-	// as well as their go.mod files.
-	//
-	// Remote dependencies are expected to be resolved from main packages'
-	// go.mod and local dependencies' go.mod files, which all must be in
-	// the tree.
-	localModules, err := localModules(pkgDir, bbDir, mainPkgs)
-	if err != nil {
+	// TODO(chrisko): We don't have to do this anymore.
+	for _, cg := range f.Comments {
+		for _, c := range cg.List {
+			if strings.HasPrefix(c.Text, "// import") {
+				c.Text = "// Import" + c.Text[9:]
+			}
+		}
+	}
+	return hasMain
+}
+
+// WritePkg writes p's files into destDir.
+func WritePkg(p *packages.Package, destDir string) error {
+	if err := os.MkdirAll(destDir, 0755); err != nil {
 		return err
 	}
 
-	var localDepPkgs []*packages.Package
-	for _, p := range mainPkgs {
-		// Find all dependency packages that are *within* module boundaries for this package.
-		//
-		// writeDeps also copies the go.mod into the right place.
-		localDeps := collectDeps(p.Pkg, localModules)
-		localDepPkgs = append(localDepPkgs, localDeps...)
-	}
-
-	// TODO(chrisko): We need to go through mainPkgs Module definitions to
-	// find exclude and replace directives, which only have an effect in
-	// the main module's go.mod, which will be the top-level go.mod we
-	// write.
-	//
-	// mainPkgs module files expect to be "the main module", since those
-	// are where Go compilation would normally occur.
-	//
-	// The top-level go.mod must have copies of the mainPkgs' modules'
-	// replace and exclude directives. If they conflict, we need to have a
-	// legible error message for the user.
-
-	// Copy local dependency packages into module directories at
-	// tmpDir/src.
-	seenIDs := make(map[string]struct{})
-	for _, p := range localDepPkgs {
-		if _, ok := seenIDs[p.ID]; !ok {
-			if err := bbinternal.WritePkg(p, filepath.Join(pkgDir, p.PkgPath)); err != nil {
-				return fmt.Errorf("writing package %s failed: %v", p, err)
-			}
-			seenIDs[p.ID] = struct{}{}
+	for _, fp := range p.OtherFiles {
+		if err := cp.Copy(fp, filepath.Join(destDir, filepath.Base(fp))); err != nil {
+			return fmt.Errorf("copy failed: %v", err)
 		}
 	}
 
-	// Avoid go.mod in the case of GO111MODULE=(auto|off) if there are no modules.
-	if env.GO111MODULE == "on" || len(localModules) > 0 {
-		// go.mod for the bb binary.
-		//
-		// Add local replace rules for all modules we're compiling.
-		//
-		// This is the only way to locally reference another modules'
-		// repository. Otherwise, go'll try to go online to get the source.
-		//
-		// The module name is something that'll never be online, lest Go
-		// decides to go on the internet.
-		content := `module bb.u-root.com/bb`
-		for mpath, module := range localModules {
-			v := module.Version
-			if len(v) == 0 {
-				// When we don't know the version, we can plug
-				// in a "Go-generated" version number to get
-				// past the validation in the compiler.
-				//
-				// We don't always do this because if the
-				// module path has a /v2 or /v3, Go expects the
-				// version number to match. So we use the
-				// module.Version when available, because it's
-				// the most accurate thing.
-				v = "v0.0.0"
-			}
-			content += fmt.Sprintf("\nrequire %s %s\n", mpath, v)
-			content += fmt.Sprintf("replace %s => ../../%s\n", mpath, mpath)
-		}
+	return writeFiles(destDir, p.Fset, p.Syntax)
+}
 
-		// TODO(chrisko): add other go.mod files' replace and exclude
-		// directives.
-		//
-		// Warn the user if they are potentially incompatible.
-		if err := ioutil.WriteFile(filepath.Join(bbDir, "go.mod"), []byte(content), 0755); err != nil {
+func writeFiles(destDir string, fset *token.FileSet, files []*ast.File) error {
+	// Write all files out.
+	for _, file := range files {
+		name := fset.File(file.Package).Name()
+
+		path := filepath.Join(destDir, filepath.Base(name))
+		if err := writeFile(path, fset, file); err != nil {
 			return err
 		}
-		return nil
 	}
 	return nil
 }
 
-func versionNum(mpath string) string {
-	last := path.Base(mpath)
-	if len(last) == 0 {
-		return "v0"
+// Rewrite rewrites p into destDir as a bb package, rewriting its init and main
+// functions.
+//
+// bbImportPath is the importpath to use for bbmain. bbImportPath is usually
+// bb.u-root.com/bb/pkg/bbmain for the Go module/vendor-based compilations, but
+// github.com/u-root/gobusybox/src/pkg/bb/bbmain for bazel-based compilations.
+func (p *Package) Rewrite(destDir, bbImportPath string) error {
+	// This init holds all variable initializations.
+	//
+	// func init0() {}
+	varInit := &ast.FuncDecl{
+		Name: p.nextInit(true),
+		Type: &ast.FuncType{
+			Params:  &ast.FieldList{},
+			Results: nil,
+		},
+		Body: &ast.BlockStmt{},
 	}
-	if matched, _ := regexp.Match("v[0-9]+", []byte(last)); matched {
-		return last
-	}
-	return "v0"
-}
 
-// deps recursively iterates through imports and returns the set of packages
-// for which filter returns true.
-func deps(p *packages.Package, filter func(p *packages.Package) bool) []*packages.Package {
-	var pkgs []*packages.Package
-	packages.Visit([]*packages.Package{p}, nil, func(pkg *packages.Package) {
-		if filter(pkg) {
-			pkgs = append(pkgs, pkg)
+	var mainFile *ast.File
+	for _, sourceFile := range p.Pkg.Syntax {
+		if hasMainFile := p.rewriteFile(sourceFile); hasMainFile {
+			mainFile = sourceFile
 		}
-	})
-	return pkgs
-}
-
-func locallyReplacedModules(p *packages.Package) map[string]*packages.Module {
-	if p.Module == nil {
-		return nil
+	}
+	if mainFile == nil {
+		return fmt.Errorf("no main function found in package %q", p.Pkg.PkgPath)
 	}
 
-	m := make(map[string]*packages.Module)
-	// Collect all "local" dependency packages that are in `replace`
-	// directives somewhere, to be copied into the temporary directory
-	// structure later.
-	packages.Visit([]*packages.Package{p}, nil, func(pkg *packages.Package) {
-		if pkg.Module != nil && pkg.Module.Replace != nil && isReplacedModuleLocal(pkg.Module.Replace) {
-			m[pkg.Module.Path] = pkg.Module
+	// Add variable initializations to Init0 in the right order.
+	for _, initStmt := range p.Pkg.TypesInfo.InitOrder {
+		a, ok := p.initAssigns[initStmt.Rhs]
+		if !ok {
+			return fmt.Errorf("couldn't find init assignment %s", initStmt)
 		}
-	})
-	return m
-}
-
-func collectDeps(p *packages.Package, localModules map[string]*packages.Module) []*packages.Package {
-	if p.Module != nil {
-		// Collect all "local" dependency packages, to be copied into
-		// the temporary directory structure later.
-		return deps(p, func(pkg *packages.Package) bool {
-			// Replaced modules can be local on the file system.
-			if pkg.Module != nil && pkg.Module.Replace != nil && isReplacedModuleLocal(pkg.Module.Replace) {
-				return true
-			}
-
-			// Is this a dependency within a local module?
-			for modulePath := range localModules {
-				if strings.HasPrefix(pkg.PkgPath, modulePath) {
-					return true
-				}
-			}
-			return false
-		})
+		varInit.Body.List = append(varInit.Body.List, a)
 	}
 
-	// If modules are not enabled, we need a copy of *ALL*
-	// non-standard-library dependencies in the temporary directory.
-	return deps(p, func(pkg *packages.Package) bool {
-		// First component of package path contains a "."?
-		//
-		// Poor man's standard library test.
-		firstComp := strings.SplitN(pkg.PkgPath, "/", 2)
-		return strings.Contains(firstComp[0], ".")
-	})
+	// import bbmain "bbImportPath"
+	importName := p.newImportName("bbmain", mainFile)
+	astutil.AddNamedImport(p.Pkg.Fset, mainFile, importName, bbImportPath)
+
+	// func init() {
+	//   bbmain.Register("p.name", Init, Main)
+	// }
+	bbRegisterSelf := &ast.FuncDecl{
+		Name: ast.NewIdent("init"),
+		Type: &ast.FuncType{},
+		Body: &ast.BlockStmt{
+			List: []ast.Stmt{
+				&ast.ExprStmt{X: &ast.CallExpr{
+					Fun: ast.NewIdent(fmt.Sprintf("%s.Register", importName)),
+					Args: []ast.Expr{
+						// name=
+						&ast.BasicLit{
+							Kind:  token.STRING,
+							Value: strconv.Quote(p.Name),
+						},
+						// init=
+						ast.NewIdent(p.init.Name.Name),
+						// main=
+						ast.NewIdent(p.mainFuncName),
+					},
+				}},
+			},
+		},
+	}
+
+	mainFile.Decls = append(mainFile.Decls, varInit, p.init, bbRegisterSelf)
+
+	return WritePkg(p.Pkg, destDir)
+}
+
+func writeFile(path string, fset *token.FileSet, f *ast.File) error {
+	var buf bytes.Buffer
+	if err := format.Node(&buf, fset, f); err != nil {
+		return fmt.Errorf("error formatting Go file %q: %v", path, err)
+	}
+	return writeGoFile(path, buf.Bytes())
+}
+
+func writeGoFile(path string, code []byte) error {
+	// Format the file. Do not fix up imports, as we only moved code around
+	// within files.
+	opts := imports.Options{
+		Comments:   true,
+		TabIndent:  true,
+		TabWidth:   8,
+		FormatOnly: true,
+	}
+	code, err := imports.Process("commandline", code, &opts)
+	if err != nil {
+		return fmt.Errorf("bad parse while processing imports %q: %v", path, err)
+	}
+
+	if err := ioutil.WriteFile(path, code, 0644); err != nil {
+		return fmt.Errorf("error writing Go file to %q: %v", path, err)
+	}
+	return nil
 }
