+++ src/pkg/monoimporter/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/u-root/gobusybox/src/pkg/monoimporter",
     visibility = ["//visibility:public"],
     deps = [
-        "//src/pkg/bb/bbinternal",
+        "//src/pkg/bb",
         "@org_golang_x_tools//go/gcexportdata",
         "@org_golang_x_tools//go/packages",
     ],
