+++ src/pkg/stdgobb/BUILD.bazel
@@ -1,0 +1,22 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_embed_data", "go_library")
+
+go_library(
+    name = "bb",
+    srcs = [
+        "bb.go",
+        "bbmain_src.go",
+        "bbregister_src.go",
+        "generate.go",
+    ],
+    importpath = "github.com/u-root/gobusybox/src/pkg/stdgobb",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/pkg/bb",
+        "//src/pkg/golang",
+        "//src/pkg/stdgobb/bbpkg",
+        "@com_github_google_goterm//term",
+        "@com_github_u_root_uio//cp",
+        "@org_golang_x_tools//go/ast/astutil",
+        "@org_golang_x_tools//go/packages",
+    ],
+)
